Implement GTA control scheme ( WASD to move character, mouse to control heading of avatar and camera, third person ). Add a hotkey to open the chat input ( you need the keyboard to move so it can't just start typing ). Update the header for numerous files. Add methods for muting music and sfx in CIBase. Add an optional config variable to toggle loading things during the loading screen. Add a fading effect for the credits in the book. Change the throw-gag button to Mouse1 ( for GTA controls ). Fix district crash relating to Suits. Disable all of the other theme songs, just play the original theme. Start working on an overhaul for the unfinished delivery minigame. DistributedMinigame on the client side now knows how many players there are. Toons now leave for the minigame sooner after teleporting away. Fix mouse sensitivity slowing down with frame rate. Fix floating snowballs in Dodgeball minigame. Avoid lighting and shading on chat balloons and nametags. Fix quest crash. Implement better animations for moving diagonally and horizontally in GTA controls. Only use the two front view camera angles if we're not using GTA controls. Update the IP for the deployment tool. Update crash minigame prototype. Update settings. Update server and client configs. Update header generator. Implement lighting and shading on the scene. Different playgrounds can have different configurations of lighting. Add materials to certain objects (phong on cogs, toons, shiny things, etc). Remove downscaling of movement speed when strafing. Convert any grayscale textures to RGB format upon Actor/Model loading. Possibly some stuff that I missed.
Former-commit-id: 653260a7e3bf7f0265c76638036e14d7e1d768f7

+++ game/src/coginvasion/minigame/DistributedMinigame.py
@@ -89,6 +89,7 @@ class DistributedMinigame(DistributedObject.DistributedObject, Timer.Timer):
         self.winnerPrize = 0
         self.loserPrize = 0
         self.round = 0
+        self.numPlayers = 0
         self.winnerMsg = "Winner!\nYou have earned: %s Jellybeans"
         self.loserMsg = "Loser!\nYou have earned: %s Jellybeans"
         self.allWinnerMsgs = ["Nice try!\nYou have earned: %s", "Good job!\nYou have earned: %s",
@@ -102,6 +103,12 @@ class DistributedMinigame(DistributedObject.DistributedObject, Timer.Timer):
         self.gameOverLbl.setBin('gui-popup', 60)
         self.gameOverLbl.hide()
         return
+
+    def setNumPlayers(self, num):
+        self.numPlayers = num
+
+    def getNumPlayers(self):
+        return self.numPlayers
 
     def roundOver(self):
         pass
@@ -322,19 +329,24 @@ class DistributedMinigame(DistributedObject.DistributedObject, Timer.Timer):
     def disable(self):
         self.deleteTimer()
         base.localAvatar.getGeomNode().setColorScale(VBase4(1, 1, 1, 1))
-        self.gameOverLbl.destroy()
-        self.gameOverLbl = None
+        if hasattr(self, 'gameOverLbl') and self.gameOverLbl is not None:
+            self.gameOverLbl.destroy()
+            self.gameOverLbl = None
         NametagGlobals.setWant2dNametags(True)
         base.localAvatar.setPosHpr(0, 0, 0, 0, 0, 0)
-        self.fsm.requestFinalState()
-        del self.fsm
+        if hasattr(self, 'fsm'):
+            self.fsm.requestFinalState()
+            del self.fsm
         self.winSfx = None
         self.loseSfx = None
         self.prizeHigh = None
         self.prizeLow = None
-        self.headPanels.delete()
-        self.headPanels = None
-        self.finalScoreUI.unload()
-        self.finalScoreUI = None
+        if self.headPanels is not None:
+            self.headPanels.delete()
+            self.headPanels = None
+        if self.finalScoreUI is not None:
+            self.finalScoreUI.unload()
+            self.finalScoreUI = None
+        self.numPlayers = None
         base.minigame = None
         DistributedObject.DistributedObject.disable(self)
